Move TimesStamp type from tuf.h to types.h
We'd like to use it in a lot of unrelated contexts
